Remove redundant methods
